move store_url.py to shared utilities (#2642)
