+++ assets/model_monitoring/components/src/model_data_collector_preprocessor/genai_run.py
@@ -13,7 +13,7 @@ from pyspark.sql.functions import lit
 from pyspark.sql.utils import AnalysisException
 from shared_utilities.df_utils import try_get_df_column
 from shared_utilities.io_utils import save_spark_df_as_mltable
-from model_data_collector_preprocessor.store_url import StoreUrl
+from shared_utilities.store_url import StoreUrl
 
 from model_data_collector_preprocessor.mdc_utils import (
     _mdc_uri_folder_to_preprocessed_spark_df,
